# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LuraDB API",
    "description": "REST-native multi-model database — KeyValue- und JSON-Engine. `version` ist die API-Contract-Version (siehe API_VERSION), unabhängig von der Server-Version; letztere steht in der Extension `x-luradb-server-version`. Kompatibilitäts-Ranges: api/COMPATIBILITY.md. Laufzeit-Check: GET /version.",
    "contact": {
      "name": "Heiko Wein"
    },
    "license": {
      "name": "FSL-1.1-ALv2"
    },
    "version": "0.1.0",
    "x-luradb-server-version": "0.1.0"
  },
  "paths": {
...
    "/store-api/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Returns all active domains with their names and creation timestamps.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DomainResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Creates a new logical domain (isolated KV namespace).",
        "description": "Domain names must be unique, max 50 chars, and match `[a-zA-Z0-9_-]`.",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          }
        }
      }
    },
    "/store-api/domains/{name}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Returns metadata for a single domain by name. Returns 404 if it does not exist.",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Schedules a domain and all its data for deletion.",
        "description": "The domain becomes immediately inaccessible; the actual purge runs in the background.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge started"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      }
    },
    "/store-api/json/domains": {
      "get": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Returns all active JSON domains.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active JSON domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JsonDomainResponse"
                  }
                }
              }
            }
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      },
      "post": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Creates a new JSON domain (isolated document namespace, separate from KV domains).",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJsonDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonDomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      }
    },
...
    "/store-api/kv/{domain}/keys": {
      "get": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Lists all keys in the domain, optionally filtered by a prefix. Returns an empty array if no keys match.",
        "operationId": "scan_keys",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "description": "Key prefix filter",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "410": {
            "description": "Domain is being deleted"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      }
    },
    "/store-api/kv/{domain}/keys/{key}": {
      "get": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Retrieves the raw byte value for a key. Returns 404 if the key does not exist or has expired.",
        "operationId": "get_key",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Key (valid UTF-8, max 256 bytes)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Value (raw bytes)"
          },
          "400": {
            "description": "Invalid key"
          },
          "404": {
            "description": "Key or domain not found"
          },
          "410": {
            "description": "Domain is being deleted"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Inserts or updates a value for the given key (upsert semantics).",
        "description": "The optional `ttl` query parameter sets an expiry in seconds; omitting it stores the key indefinitely.",
        "operationId": "put_key",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Key (valid UTF-8, max 256 bytes)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ttl",
            "in": "query",
            "description": "TTL in seconds (optional)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Key upserted"
          },
          "400": {
            "description": "Invalid key"
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is being deleted"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Permanently removes a key from the domain. The operation is idempotent; deleting a non-existent key still returns 204.",
        "operationId": "delete_key",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Key (valid UTF-8, max 256 bytes)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key deleted"
          },
          "400": {
            "description": "Invalid key"
          },
          "410": {
            "description": "Domain is being deleted"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      }
    },
    "/store-api/kv/{domain}/keys/{key}/null": {
      "patch": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Writes an explicit null/tombstone marker for a key without removing it from the keyspace.",
        "description": "Useful for signalling soft-deletes in distributed or CDC scenarios.",
        "operationId": "set_null",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Key (valid UTF-8, max 256 bytes)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Key set to null (tombstone)"
          },
          "400": {
            "description": "Invalid key"
          },
          "410": {
            "description": "Domain is being deleted"
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      }
    },
    "/store-api/kv/{domain}/watch": {
      "get": {
        "tags": [
          "Key-Value Store"
        ],
        "summary": "Opens a Server-Sent Events stream that delivers real-time change events (`set` / `delete`) for keys matching the optional prefix.",
        "description": "The connection stays open until the client disconnects; a keep-alive ping is sent automatically.",
        "operationId": "watch",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "description": "Key prefix filter",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of key change events"
          },
          "410": {
            "description": "Domain is being deleted"
          }
        }
      }
    },
...
    "/store-api/rel/domains": {
      "get": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Returns all active relational domains.",
        "operationId": "list_domains",
        "responses": {
          "200": {
            "description": "List of active relational domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelDomainResponse"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      },
      "post": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Creates a new relational domain (isolated table/view namespace, separate",
        "description": "from KV and JSON domains).",
        "operationId": "create_domain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Relational domain created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelDomainResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain name"
          },
          "409": {
            "description": "Domain already exists"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDomainRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "User-visible domain name (max 50 chars, [a-zA-Z0-9_-])."
          }
        }
      },
...
      "DomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "User-visible domain name."
          }
        }
      },
...
      "JsonDomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "document_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of documents — only set on the detail endpoint.",
            "nullable": true,
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Lifecycle state: \"active\" or \"deleting\" (background purge running)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key wie in `luradb.toml` konfiguriert (Admins) oder wie bei User-Anlage / Key-Rotation erhalten. Nur bei `auth.enabled = true` erforderlich."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}